+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - only_preferred_wallet option support
 - added dispenser edit page
 - updates for /dispensers endpoints
+- new events for Plausible
 
 ## 2.1.12
 - Added Dispensers page for whitelisted users (Multi QR)